requirements.txt
django
cloudinary
pillow